fix: replacing any types with proper types

+++ src/components/client-components/DeleteButton.tsx
@@ -9,7 +9,9 @@ const DeleteButton = ({ userId, dataId }: { userId?: string, dataId?: string }) 
     <DropdownMenuItem className="hover:bg-white/5 rounded-xl text-red-600 cursor-pointer px-4 py-1 mt-2 text-sm" onClick={(e) => e.stopPropagation()}>
       <div onClick={(e) => {
         e.stopPropagation()
-        handleDeletePost({ userId, dataId })
+        if(userId && dataId) {
+          handleDeletePost({ userId, dataId })
+        }
       }}>Delete</div>
     </DropdownMenuItem>
